# artifacts/api-server/src/routes/auth.ts
import { Router } from "express";
import { db } from "@workspace/db";
import { usersTable } from "@workspace/db/schema";
import { eq } from "drizzle-orm";
import { hash, compare } from "bcryptjs";
import { signToken } from "../lib/jwt";

const router = Router();

router.post("/auth/login", async (req, res) => {
  try {
    const { email, password } = req.body;
    if (!email || !password) {
      return res.status(400).json({ error: "Email and password are required" });
    }
    const [user] = await db.select().from(usersTable).where(eq(usersTable.email, email));
    if (!user) {
      return res.status(401).json({ error: "Invalid credentials" });
    }
    const valid = await compare(password, user.password || "");
    if (!valid) {
      return res.status(401).json({ error: "Invalid credentials" });
    }
    const token = signToken(user.id);
    return res.json({ token, user: { id: user.id, name: user.name, email: user.email, role: user.role, systemRole: user.systemRole } });
  } catch (err) {
    console.error("Login error:", err);
    return res.status(500).json({ error: "Internal error" });
  }
});

export default router;