# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AgencyOS API
servers:
  - url: /api
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
  /invoices/{id}:
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List proposals
      responses:
        "200":
          description: Proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Create proposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}:
    patch:
      operationId: updateProposal
      tags: [proposals]
      summary: Update proposal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
    delete:
      operationId: deleteProposal
      tags: [proposals]
      summary: Delete proposal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /leaves:
    get:
      operationId: listLeaveRequests
      tags: [leaves]
      summary: List leave requests
      parameters:
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeaveRequest
      tags: [leaves]
      summary: Submit a leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}/approve:
    post:
      operationId: approveLeaveRequest
      tags: [leaves]
      summary: Approve a leave request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}/reject:
    post:
      operationId: rejectLeaveRequest
      tags: [leaves]
      summary: Reject a leave request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
...
  /proforma-invoices:
    get:
      operationId: listProformaInvoices
      tags: [proforma-invoices]
      summary: List proforma invoices
      responses:
        "200":
          description: List of proforma invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProformaInvoice"
    post:
      operationId: createProformaInvoice
      tags: [proforma-invoices]
      summary: Create proforma invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProformaInvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProformaInvoice"
  /proforma-invoices/{id}:
    get:
      operationId: getProformaInvoice
      tags: [proforma-invoices]
      summary: Get proforma invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProformaInvoice"
    patch:
      operationId: updateProformaInvoice
      tags: [proforma-invoices]
      summary: Update proforma invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProformaInvoiceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProformaInvoice"
    delete:
      operationId: deleteProformaInvoice
      tags: [proforma-invoices]
      summary: Delete proforma invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
...
    User:
      type: object
      required: [id, name, email, systemRole]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        systemRole:
          type: string
        department:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
...
    Invoice:
      type: object
      required: [id]
      properties:
        id:
          type: string
        number:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        subtotal:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        total:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        companyGstin:
          type: ["string", "null"]
        clientGstin:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        shippingAddress:
          type: ["string", "null"]
        termsAndConditions:
          type: ["string", "null"]
        bankDetails:
          type: ["object", "null"]
        lineItems:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/LineItem"
    InvoiceInput:
      type: object
      properties:
        clientId:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        subtotal:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        total:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        companyGstin:
          type: ["string", "null"]
        clientGstin:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        shippingAddress:
          type: ["string", "null"]
        termsAndConditions:
          type: ["string", "null"]
        bankDetails:
          type: ["object", "null"]
        lineItems:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/LineItem"
    LineItem:
      type: object
      properties:
        description:
          type: string
        hsnSac:
          type: ["string", "null"]
        qty:
          type: number
        unitPrice:
          type: number
        taxPercent:
          type: number
...
    Proposal:
      type: object
      required: [id]
      properties:
        id:
          type: string
        title:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        template:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        validUntil:
          type: ["string", "null"]
        scope:
          type: ["string", "null"]
        deliverables:
          type: ["string", "null"]
        timeline:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ProposalInput:
      type: object
      properties:
        title:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        template:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        validUntil:
          type: ["string", "null"]
        scope:
          type: ["string", "null"]
        deliverables:
          type: ["string", "null"]
        timeline:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    LeaveRequest:
      type: object
      required: [id, userId, type, startDate, endDate, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: ["string", "null"]
        type:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    LeaveRequestInput:
      type: object
      required: [type, startDate, endDate]
      properties:
        type:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
...
    ProformaInvoice:
      type: object
      required: [id]
      properties:
        id:
          type: string
        number:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        subtotal:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        total:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        companyGstin:
          type: ["string", "null"]
        clientGstin:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        shippingAddress:
          type: ["string", "null"]
        termsAndConditions:
          type: ["string", "null"]
        bankDetails:
          type: ["object", "null"]
        lineItems:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/LineItem"
        createdAt:
          type: string
    ProformaInvoiceInput:
      type: object
      properties:
        clientId:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        subtotal:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        total:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        companyGstin:
          type: ["string", "null"]
        clientGstin:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        shippingAddress:
          type: ["string", "null"]
        termsAndConditions:
          type: ["string", "null"]
        bankDetails:
          type: ["object", "null"]
        lineItems:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/LineItem"
...