# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AgencyOS API
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceBankDetails } from './invoiceBankDetails';
import type { LineItem } from './lineItem';

export interface Invoice {
  id: string;
  /** @nullable */
  number?: string | null;
  /** @nullable */
  clientId?: string | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  invoiceDate?: string | null;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  subtotal?: number | null;
  /** @nullable */
  taxAmount?: number | null;
  /** @nullable */
  total?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  companyGstin?: string | null;
  /** @nullable */
  clientGstin?: string | null;
  /** @nullable */
  billingAddress?: string | null;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  termsAndConditions?: string | null;
  /** @nullable */
  bankDetails?: InvoiceBankDetails;
  /** @nullable */
  lineItems?: LineItem[] | null;
}